// src/app/interfaces/user-data.ts
export interface UserData {
    username: string;
    image: string;
    type: string;
    availabilityTime: number;
    availabilityTimeDuration: string;
    supplierTimeService: string;
    supplierTime: number;
    repetitiveReservation: number;
    repetitiveReservationType: string;
    calendar : any
  }

  